requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
SQLAlchemy==2.0.23
openpyxl==3.1.2
python-dotenv==1.0.0
WTForms==3.1.1
email-validator==2.1.0.post1
Flask-Migrate==4.0.4
Flask-Script==2.0.6

